+++ src/lib.rs
@@ -8,7 +8,7 @@ mod storage;
 mod tcp_io_parser;
 
 use futures::{SinkExt as _, StreamExt as _};
-use log::{error, info};
+use log::{error, info, warn};
 use std::sync::Arc;
 use tokio::net::{TcpListener, TcpStream};
 use tokio::sync::Semaphore;
@@ -17,8 +17,9 @@ use tokio_util::codec::Decoder as _;
 use crate::background_merge::BackgroundMerge;
 use crate::config::CONFIG;
 use crate::error::Error;
+use crate::runtime_config::{TABLE_DATA, TableConfig};
 use crate::sql::CommandRunner;
-use crate::storage::load_all_parts_on_startup;
+use crate::storage::{TableDef, TableMetadata, TablePartInfo};
 use crate::tcp_io_parser::Parser;
 
 pub fn reject_32bit_systems() -> Result<(), String> {
@@ -36,11 +37,6 @@ pub fn build_logger() {
     env_logger::Builder::from_default_env()
         .filter_level(CONFIG.get_log_level())
         .init();
-}
-
-pub fn load_parts() -> Result<(), String> {
-    load_all_parts_on_startup(CONFIG.get_db_dir())
-        .map_err(|error| format!("Failed to load parts on startup: {error:?}"))
 }
 
 pub fn spawn_background_merges() {
@@ -130,5 +126,136 @@ async fn handle_connection(socket: &mut TcpStream) -> Result<(), Error> {
         }
     }
     info!("Connection closed.");
+    Ok(())
+}
+
+/// Loads all table parts from filesystem into memory on startup.
+///
+/// Scans all databases and tables, loads part indexes, and populates `TABLE_DATA`.
+/// Cleans up any leftover raw directories from previous runs.
+///
+/// Returns: Ok or `String` on critical failure
+pub fn load_all_parts_on_startup() -> Result<(), String> {
+    let db_dir = CONFIG.get_db_dir();
+    info!(
+        "Loading parts from database directory: {}",
+        db_dir.display()
+    );
+
+    if !db_dir.exists() {
+        warn!("Database directory does not exist: {}", db_dir.display());
+        return Err(format!(
+            "Database directory does not exist: {}",
+            db_dir.display()
+        ));
+    }
+
+    let databases = std::fs::read_dir(db_dir).map_err(|error| {
+        format!(
+            "Failed to read database directory ({}): {error}",
+            db_dir.display()
+        )
+    })?;
+
+    for database_entry in databases {
+        let database_entry =
+            database_entry.map_err(|error| format!("Failed to read database entry: {error}"))?;
+
+        let database_path = database_entry.path();
+        if !database_path.is_dir() {
+            continue;
+        }
+
+        let database_name = database_entry.file_name().to_string_lossy().to_string();
+        let tables = std::fs::read_dir(&database_path).map_err(|error| {
+            format!("Failed to read tables in database {database_name:?}: {error}")
+        })?;
+
+        for table_entry in tables {
+            let table_entry =
+                table_entry.map_err(|error| format!("Failed to read table entry: {error}"))?;
+
+            let table_path = table_entry.path();
+            if !table_path.is_dir() {
+                continue;
+            }
+
+            let table = table_entry.file_name().to_string_lossy().to_string();
+            let table_def = TableDef {
+                database: database_name.clone(),
+                table,
+            };
+
+            let table_metadata = TableMetadata::read_from(&table_def).map_err(|error| {
+                format!(
+                    "Could not read table metadata from table definition ({table_def}): {error}"
+                )
+            })?;
+
+            TABLE_DATA.insert(
+                table_def.clone(),
+                TableConfig {
+                    metadata: table_metadata,
+                    infos: Vec::new(),
+                },
+            );
+
+            let parts = std::fs::read_dir(&table_path).map_err(|error| {
+                format!("Failed to read parts from table definition ({table_def}): {error}")
+            })?;
+
+            for part_entry in parts {
+                let part_entry = part_entry.map_err(|error| {
+                    format!("Failed to read part entry for table definition ({table_def}): {error}")
+                })?;
+
+                let part_path = part_entry.path();
+                let part_name = part_entry.file_name().to_string_lossy().to_string();
+
+                if !part_path.is_dir() || part_name.starts_with('.') {
+                    continue;
+                }
+
+                if part_name == "raw" {
+                    match std::fs::remove_dir_all(&part_path) {
+                        Ok(()) => {
+                            info!("Removed raw directory for table {table_def}");
+                        }
+                        Err(e) => {
+                            warn!("Failed to remove raw directory for table {table_def}: {e}");
+                        }
+                    }
+                    continue;
+                }
+
+                if part_path
+                    .extension()
+                    .is_some_and(|ext| ext.eq_ignore_ascii_case("old"))
+                {
+                    warn!(
+                        "Found old part: {part_name}. Consult the logs to make the decision about removal."
+                    );
+                    continue;
+                }
+
+                match TablePartInfo::read_from(&table_def, &part_name) {
+                    Ok(info) => {
+                        let Some(mut result) = TABLE_DATA.get_mut(&table_def) else {
+                            panic!(
+                                "Table definiton is removed from TABLE_DATA while loading all parts.."
+                            ) // todo: why would it return an Option??
+                        };
+                        result.infos.push(info);
+                        info!("Loaded part ({part_name}) for table ({table_def})");
+                    }
+                    Err(error) => {
+                        warn!("Failed to load part {part_name} for table {table_def}: {error:?}");
+                    }
+                }
+            }
+        }
+    }
+
+    info!("Finished loading parts");
     Ok(())
 }
